Add and use tuple apply
std::apply is part of c++17, and is therefore not as widely supported.
However, some standard libraries do support it. ADL then activates, as
the second argument is std::tuple in namespace std, also including
std::apply in the overload set. This is why it needs to have the
namespace specified in some places.